Common abstract class for attributes adapter
