+++ network-store-iidm-impl/src/main/java/com/powsybl/network/store/iidm/impl/BusbarSectionToInjectionAdapter.java
@@ -6,60 +6,16 @@
  */
 package com.powsybl.network.store.iidm.impl;
 
-import com.powsybl.network.store.model.*;
-
-import java.util.Map;
-import java.util.Set;
+import com.powsybl.network.store.model.BusbarSectionAttributes;
+import com.powsybl.network.store.model.ConnectablePositionAttributes;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class BusbarSectionToInjectionAdapter implements InjectionAttributes {
-
-    private final BusbarSectionAttributes attributes;
+public class BusbarSectionToInjectionAdapter extends AbstractIdentifiableToInjectionAttributesAdapter<BusbarSectionAttributes> {
 
     public BusbarSectionToInjectionAdapter(BusbarSectionAttributes attributes) {
-        this.attributes = attributes;
-    }
-
-    @Override
-    public Resource getResource() {
-        return attributes.getResource();
-    }
-
-    @Override
-    public void setResource(Resource resource) {
-        attributes.setResource(resource);
-    }
-
-    @Override
-    public String getName() {
-        return attributes.getName();
-    }
-
-    @Override
-    public void setName(String name) {
-        attributes.setName(name);
-    }
-
-    @Override
-    public boolean isFictitious() {
-        return attributes.isFictitious();
-    }
-
-    @Override
-    public void setFictitious(boolean fictitious) {
-        attributes.setFictitious(fictitious);
-    }
-
-    @Override
-    public Map<String, String> getProperties() {
-        return attributes.getProperties();
-    }
-
-    @Override
-    public void setProperties(Map<String, String> properties) {
-        attributes.setProperties(properties);
+        super(attributes);
     }
 
     @Override
@@ -130,35 +86,5 @@ public class BusbarSectionToInjectionAdapter implements InjectionAttributes {
     @Override
     public void setPosition(ConnectablePositionAttributes position) {
         throw new AssertionError();
-    }
-
-    @Override
-    public Set<String> getAliasesWithoutType() {
-        return attributes.getAliasesWithoutType();
-    }
-
-    @Override
-    public void setAliasesWithoutType(Set<String> aliasesWothoutType) {
-        attributes.setAliasesWithoutType(aliasesWothoutType);
-    }
-
-    @Override
-    public Map<String, String> getAliasByType() {
-        return attributes.getAliasByType();
-    }
-
-    @Override
-    public void setAliasByType(Map<String, String> aliasByType) {
-        attributes.setAliasByType(aliasByType);
-    }
-
-    @Override
-    public Map<String, ExtensionAttributes> getExtensionAttributes() {
-        return attributes.getExtensionAttributes();
-    }
-
-    @Override
-    public void setExtensionAttributes(Map<String, ExtensionAttributes> extensionAttributes) {
-        attributes.setExtensionAttributes(extensionAttributes);
     }
 }
